statsbot/requirements.txt
discord
discord.py
aiohttp
beautifulsoup4
html2image
Pillow
